[roll] Roll fuchsia [io][conformance] Make ulibfs a child component in conformance test
Deletes the io_fidl_conformance_tests package, which used to hardcode
the ulibfs test harness as the only harness it tested, and replaces it
with a new test package specifically for ulibfs. This construction can
be easily replicated for other filesystems in the future.

This also significantly simplifies the handshake between Io1Harness and
Io1HarnessReceiver. In fact, the receiver is not needed anymore at all,
so that has been deleted.

Original-Bug: 33880
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/422456
Original-Revision: 56a7056db37d84fec2a1c5ad7eb682c2c1979ca2
GitOrigin-RevId: 6e996f5f33f74bb74bef7ca5fb6d2911011b9229
Change-Id: I6009ed79c5818c1f39e4f1a27cece5edbadc7e82