[roll] Roll cobalt [cobalt 1.1 privacy] Check in initial table of privacy encoding params
This table was generated by pasin@ and will be used by the
registry parser to look up the privacy encoding parameters for
report with added privacy. This is just an initial version of the
table for use in development.

The column format is:
{epsilon, population, sparsity, prob_bit_flip, num_index_points}

where:
* epsilon is the target (shuffled) epsilon value,
* population is the estimated size of the fleet,
* sparsity is the maximum number of buckets with nonzero counts in
  the histogram contributed by an individual client.
* See the ReportDefinition proto for the meanings of the
  prob_bit_flip and num_index_points fields.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt/+/417859
Original-Revision: ade148a06c9cc19940145de9229c94b5eabe57be
GitOrigin-RevId: c33d44d0d0725ffc40e1ffb743dd254cae45663b
Change-Id: Iff8de3b0619de597216aa7a9154c660bff8a9918