[roll] Roll fuchsia [fatfs][mounter] Fix race when opening block device
If we received two requests very close together, it's possible that both
would try and open the block device. Previously the code assumed that
both would succeed, and checked that the "current" block device was
still None before overwriting it.

Unfortunately, this assumption is incorrect and what would actually
happen is that one of the requests would get back Status::ALREADY_BOUND
(which occurs when the block FIFO is being bound). Fix this by switching
to an async-friendly mutex and holding that lock while initialising the
new FatDevice.

Original-Bug: 59131
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/422538
Original-Revision: a7f207e82e52cf16303093eaaccbeb7b89e0146d
GitOrigin-RevId: b2d34f54326b8c52744ba323c8ddbdd3ea6b3d1a
Change-Id: I2384413bad9ed37cc5a57cc0dace30eca22d60c0