[roll] Roll fuchsia [camera][controller] Initial luis controller support
This CL provides initial controller support for luis.
It consists of a single configuration that provides a full-resolution
stream from the ISP.

The present form compiles and doesn't crash on a luis platform board.
Validation of the streaming is blocked on the recept of sensor frames.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/422474
Original-Revision: b33d40abc5813816ef6ba54932a50353a5b9425d
GitOrigin-RevId: 24055cd88052399fa48e20c350e5e69e9bedc1bb
Change-Id: Ie7e55745fa0c158ad20833237d8e5a617d739a60