[roll] Roll fuchsia [prebuilts] Update SwiftShader host build scripts.
(1) fxr/353570 was previously blocked by a bug in vulkan.hpp in
old Vulkan SDK. We update the SDK version to 1.2.148, and all
the dependency libraries.

(2) On some systems (e.g. Debian workstation) we saw a crash caused
by libvulkan.so.1 and libvulkan.so pointing to different files. Now
we make libvulkan.so a symbolic link to libvulkan.so.1 to avoid this.

(3) VkLayer_standard_validation is not supported by the SDK anymore.
We remove this from the build scripts.

TEST=./update-build-config.py > build_config
     ./build-linux-host-prebuilts.sh --install-dir=/tmp

Original-Bug: 43647
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/421746
Original-Revision: e8b382d89481deb802f7455416ee6db5d02a164b
GitOrigin-RevId: 7bce322c5c5da609cec04140948ef50ec5a5491f
Change-Id: I29081850b0416e659a7a8adbd86e41ec16e0ebdb