[roll] Roll fuchsia [aml][cpu] Make CPU Bind rules board specific.
We want to use a different CPU for multiple boards that have
the same SoC and it's getting clunky to match on the correct
driver.

Make the drivers bind to board specific bind rules to get
around this.

Test: Booted on Sherlock/Astro, Device Enumeration
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/422478
Original-Revision: b0019e7b1acc4801175fae741ebb5b73e9c98d6c
GitOrigin-RevId: fdf2c23220df46fcd5437ee7a9dfc3b7ef2f206c
Change-Id: I887a12fadde72554bae7cdb2f4c8f522679d9c29