[roll] Roll fuchsia [build] Consult custom_signing_script during flash.sh creation
Today we rely on a redundant `signed_image` variable to inform the
creation of the flashing script when to use signed images. (I hit a bug
in which we failed to set this for Sherlock.) This is the first of three
changes to migrate us off of `signed_image` and just consult
`custom_signing_script` directly.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/423129
Original-Revision: 28a2befb4a0ec259b38eca1a5792e3c9af02f151
GitOrigin-RevId: 1d2396d96deb04fbd220ffa9b994de5ce2e90710
Change-Id: I1280aa3bd5f7dfde2487812ca0982d7e0a17e13f