[roll] Roll fuchsia [camera][bin] Fix threading issue in FactoryServer
This CL fixes an issue whereby a callback was being passed into
RequestCapture on the factory_server thread but this was then mistakenly
invoked on the streamer thread. This lead to a crash on non-release
builds when trying to capture frames due to the `IsCurrentThreadOk()`
assert.

Test: capturing frames works correctly.
Fix: b/166159847

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/422882
Original-Revision: 74498ff951f3187ad2879e0fb085e9db648e8d46
GitOrigin-RevId: a3b3b4fdfd86ffde7c2776e52c216bebb1d43805
Change-Id: Ie2889dc189974e1809325d3a9eb47f93035a8541