[roll] Roll fuchsia [fidlc][handle] Always disallow old handle syntax
Previous changes have removed all uses of handle<> in .fidl files.
Remove the rest in tests and then always disable old-style in the
compiler. Leave the parsing of old-style for now so that an informative
error can be emitted if someone re-adds handle<> style.

Subsequently we can strip out more of the support for old-style syntax
once this has simmered for a while.

Original-Bug: 51001
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/422016
Original-Revision: b77278dc523ee8d1845ad1e238528fdaf00137ac
GitOrigin-RevId: 80ca75ce855b296756aab8cf17fd5ac613700eab
Change-Id: Iaf21a9aff6ddb458f5fc2e960ab5aa5ac53e7c69