[roll] Roll fuchsia [scenic] Fix racy annotation View crash
This CHECK will fire if the View-to-be-annotated's session is
destroyed before the AnnotationManager gets a chance to create the
annotation ViewHolder.  The dead session does not inform
AnnotationManager that the View is gone and the weak_ptr ends up
causing a CHECK.

The fix is simple: let AnnotationManager know that the View is already
gone so it can return ZX_ERR_PEER_CLOSED.  Since the Session doesn't have
access to the AnnotationManager, invalidate the entry inside of the
ViewTree to accomplish that feat indirectly.

TEST: Added tests to ViewTreeTest and AnnotationManagerTest
BUG: 58564
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414462
Original-Revision: b86b4e7bb1fae618fcba9d7f6f4317ba0b20733f
GitOrigin-RevId: 1298019148d12269b1680c18722e0fd17c297976
Change-Id: Ia5403e4628082aa105c835275e975eef93b52bd3