[roll] Roll fuchsia [netstack] Correctly implement SO_LINGER
This changes the behaviour of Close to respect SO_LINGER. man 7 socket:

  When enabled, a close(2) or shutdown(2) will not return until
  all queued messages for the socket have been successfully sent
  or the linger timeout has been reached.  Otherwise, the call
  returns immediately and the closing is done in the background.
  When the socket is closed as part of exit(2), it always
  lingers in the background.

Prior to this change, we always did the work in the background. On top
of that, we also did not correctly time out data-in-socket when the
linger time expired. This commit fixes both problems.

Unfortunately, we're back to manual reference counting.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/423274
Original-Revision: 3a15caca741f188886cd53142aee0410811ef1f6
GitOrigin-RevId: f425f35273b41caa899ba7ef926afcff9abdd6dc
Change-Id: I7a18c84543c9f719bd3a8e40bf7dc511100be4c7