[roll] Roll fuchsia [audio_core] Fix test flake in AudioCapturerPipelineTest.CaptureWithPts
Previously, this test worked as follows:

1. Compute a timestamp T such that audio_core should be reading from
   input_device.ring_buffer[0] at time T.

2. Write 10ms of data starting at input_device.ring_buffer[0].

3. Wait until now > T + 10ms

In step 1, it may happen that timestamp T is so close to now (say 1 ns
in the future) such that we have too little time to write the data
before it will be read. When we detect this scenario, we set T one full
ring buffer in the future. This introduces a problem:

            <--- ringbuff size --->
  ----+-+--+-+---------------------+-+----------
  ... |N|  |A|                     |B|    ...
  ----+-+--+-+---------------------+-+----------
       ^    ^                       ^
       |    data also               want to write
       |    readable here           data here (time T)
       |
       |<------ we capture all of these frames --->

In the above diagram, A and B are boundaries of the ring buffer (times
at which audio_core will read ring_buffer[0]) and N is the frame that
should be read at "now". We start capturing from the ring buffer at N.
We write our data at frame B so it can be read at time T. However,
because we write that data right away, it's immediately visible to the
capturer when it reads from the (equivalent) frame A.

As illustrated above, the bug is that we capture the data twice: once
during ring buffer cycle A (which should be silent) and again during
cycle B. This confuses our test which expects to see the data exactly
once.

To fix this bug, the test now works as follows:

1. Compute a ring buffer offset R such that audio_core should be reading
   from R exactly 200ms in the future.

2. Write 10ms of data starting at input_device.ring_buffer[R].

3. Wait until now > start + 200ms + 10ms

In step 2, our write might wrap around the ring buffer. I've updated
VmoBackedBuffer::WriteAt to support wrap around. Empirically on my
astro, it can take ~130ms to start a capture, hence the 200ms delay.
I don't know why it can take so long (this is a separate issue if
anything).

Multiply: fuchsia-pkg://fuchsia.com/audio-core-api-pipeline-tests#meta/audio-capturer-pipeline-test.cmx
Original-Fixed: 59351
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/423239
Original-Revision: bb1a7c879c522f724b8190ceee92cd7e46f6c995
GitOrigin-RevId: bbe9a61e2f7f05f2b5132fdd85bd949a944d46c3
Change-Id: I05e53700f35ff6419c226ba1adf1f791aba77778