[roll] Roll fuchsia [rust] Make handle invalid() a const fn
This can be called in const contexts e.g. for initializing a
struct with invalid fields or for performing branches based on
invalid / not.

In these cases, the drop function won't be called (which would
be a no-op anyways if it were).

This is not expected to have much impact on performance today.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/423400
Original-Revision: d8e2d2a8d67cebbef94e7ec8f9809ca9bb2618a1
GitOrigin-RevId: 6f29362d7e4b3a050958109782d199de66408507
Change-Id: Ie7108a56379b6bd2d9c147e7089695cbb431a495