[roll] Roll fuchsia [fint] Check in static.proto
This protobuf contains all current fuchsia.proto spec fields that are
needed by the build, and thus will eventually be consumed by fint.

I'm also checking in the Go generated code for the protobuf. I spent a
while trying to add a GN target to ensure that the generated code is
always up to date with the protobuf, but I wasn't able to get that
working, so for now we'll just have to remember to manually re-run `go
generate` after updating the proto.

Original-Bug: 59040
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/421555
Original-Revision: 8bc780fd9fbd5791b72696cc9a45df9bb62dc585
GitOrigin-RevId: c66e759e5426beb912e453d1b7c3b0b1b3bc8727
Change-Id: I44c581355a9ed3840d9dfbf021e2c9bc7a53c5c1