[roll] Roll fuchsia [a11y] Adds a semantics event callback to the semantic tree.
PLEASE READ: This change may look like it goes against what we
discusssed in regards to using callbacks vs. pointers.
The idea here is that the ViewManager will set the callback when it
instantiates a semantic tree since the ViewManager knows the correct
koid and owns the semantics event manager. This callback will populate
the view_ref_koid field of the incoming event and call
SemanticsEventManager::OnEvent. So functionally, this change is
equivalent to the semantic tree owning a pointer to the semantics event
manager, just slightly simpler to test.

TEST: fx test a11y_lib_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/417893
Original-Revision: 288fa2959d391ac827117c0b351fadcef15f3425
GitOrigin-RevId: 236372ea33841b15722787eb6a5d29715c645dce
Change-Id: I00bf174e4c2f3926e46a39a12720ab7fb0b3485e