[roll] Roll fuchsia [zbitl] Introduce storage test traits
This change simplifies the shape of the storage-specific helpers
required for the general test logic, especially in a way that will make
testing the introduction of copy/cloning storage traits easier.

We replace the state-carrying "Io" structs with simpler trait structs
of static methods. These are moved into headers in anticipation of
tests for copy traits (forthcoming), at which point we'll want to test
the cross-product of the storage types and have all of the
specializations in a single translation unit.

Also, we fix a bug in which the stdio tests were not closing their file
pointers.

Original-Bug: 57244
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/422094
Original-Revision: f66033c53e2d5a475e7cc42c189daad8e0ac5d44
GitOrigin-RevId: 7c6640e7347a63ee402ea8284519f2299bcf1b40
Change-Id: Ief66ccae9cbb35ec7d8a92c060e04c023ac14c0f