[roll] Roll fuchsia [storage] Fix rw-workers-tests.
rw-workers-test was testing files backed in memfs rather than the
filesystem under test.  This was because it forgot to append a "/" to
the mount path.

To address this, I've changed it so that the mount path always includes
a trailing slash.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/423336
Original-Revision: ae871b0530e4aa224e000e10e55cd93b08d7ba72
GitOrigin-RevId: b1fc6b4f39887f426edd4267aa7ff241f900956f
Change-Id: I666bef6377dea6560a131d6f2491e483f7b94cee