[roll] Roll fuchsia [fs-management] Recognise vbmeta partitions.
Recognise the magic header 'AVB0' of vbmeta partitions.

This is a work-around of them being misrecognised as an MBR partition if
the last two bytes of the last sector happen to be 0x55 0xAA, of which
there is a roughly 1 in 65536 chance. (These bytes tend to be ~random in
vbmeta partitions).

Original-Bug: 59374
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/423340
Original-Revision: f983b27af3931939d17f68f3fab41480c101b242
GitOrigin-RevId: 1842ca0267b7a9f7b91e2cff3d7064010f23cf71
Change-Id: I02798eaf2ca2748c7fd42dcd715253d1f0a36fe6