[roll] Roll fuchsia [flatland] Engine unittest harness.
The Flatland Engine will have two main inputs: the InstanceMap from
the UberStructSystem and the LinkTopologyMap from the LinkSystem.
Typically, these two are populated by Flatland instances, but this is
not desirable from a unittesting standpoint since it depends on a lot
of functionality from the Flatland implementation.

This harness includes the FakeFlatlandSession, which includes the
barebones necessary to create and manage the data associated with a
Flatland instance. Specifically, it provides a convenient way to set
up links and local topologies, but leaves the creation of images and
matrices to the tests themselves.

The file includes an example test that should be deleted when actual
Engine tests are being written.

TEST=fx test -o flatland_unittests
MULTIPLY: flatland_unittests (fuchsia): 30

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/422106
Original-Revision: 4c83581e944cea84e6ccfa47158936ecc4cc53a0
GitOrigin-RevId: 1ed9bea1b828e1fd6a2eef2b9cc9eeb977ae8192
Change-Id: I77b339ee5b5c16f5ef37478bacf2eea4fa0b97fc