[roll] Roll fuchsia [blobfs] Disable cache eviction
Once a blob is first opened, it will never be evicted, even after the
last handle to the blob closes.

This improves the performance of later accesses to a file, especially in
degenerate cases where a file is repeatedly reopened (which happens
during testing, particularly for test runners that spawn a new process
for each test case for hermeticity).

For example, improves the runtime of blobfs-unittest (repackaged as a V2
test component) from 4m20s to 50s, a 5x reduction in runtime.
(Subsequent local re-runs are even faster, around 20-30s).

Memory is reclaimed by the system as needed by the kernel page eviction
mechanism. Data pages that have not been recently accessed are reclaimed
when the system is under memory pressure. The only additional memory
overhead that is not reclaimed is blobfs' internal memory use for
metadata (particularly, merkle trees).

Original-Bug: 58415
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/421454
Original-Revision: 5a1c9161ad957819de0136d5e70d706e1bbed2d7
GitOrigin-RevId: a32ea3843a04736298ea82d9f5b7502168ab0483
Change-Id: Ie3d5c2dd6cbb309fdf90873f8a8844ed7983c4ad