[roll] Roll fuchsia [fidl] Specialize IsFoo<> LLCPP traits using codegen
Previously, these traits were implemented based on the
attributes of the classes themselves (e.g. checking for
a ::Type, ::MaxNumHandles, or a builder method), which
is not robust to changes to the API. Instead, this CL
emits specializations for each type in the generated code.

Test: fx build fidl_llcpp_types_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/423115
Original-Revision: b2d84fb70277045d2f937ad3a9770834a81061c0
GitOrigin-RevId: 25bcef3eafcda4280eea1095185e23b7d53f3243
Change-Id: I08d8008f1f4c5aec7ec5fbf8af3f3db206a4ccb6