[roll] Roll fuchsia [thermal] Remove MMIO offset from driver
The aml-thermal-s905d2g/-legacy drivers include a baked in 0x800 offset
that is used to address into the provided MMIO region for controlling
the sensor. Because of this, the MMIO address that is provided by the
board file has 0x800 subtracted from the true sensor base address.

It is unlcear why the offset was ever used in the first place. In its
current form, it only serves to add confusion because the MMIO address
that must be provided to the sensor driver is not part of the SoC memory
map documentation.

Test: fx test aml-thermal-s905d2g-test aml-thermal-s905d2g-legacy-test
Test: boot up Astro and ensure power_manager is still receiving
  temperature samples
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/423725
Original-Revision: deb5e2f1fabcb7e3d5549641c8b92acf67964220
GitOrigin-RevId: 265ba6e4242fc201840e8530cc6c37b4a38ae15d
Change-Id: I85abd8f8eab1b536e99f4878a5608b17700ff02f