[roll] Roll fuchsia [audio_core] Fix Gain::GetGainDb
There is additional logic in `GetGainScale` that was not being applied
in GetGainDb. Fix this by simply defining GetGainDb using GetGainScale.

This fixes an issue where we can report an incorrect stream gain/volume
to effects. This can occur because `GetGainDb` was referencing values
that are cached and only updated by `GetGainScale`. In the sampler,
there is a fast-path that checks `Gain::IsSilent`, which bypasses this
`GetGainScale` method (which is what allowed this to work before).

Test: fx test audio_core_unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/423594
Original-Revision: 8b02b3912c4bf8bfbabf2bf54e15a61eefca38f1
GitOrigin-RevId: a8d051730f542604b1aab208c76989eb22f24136
Change-Id: If947ddb759fb9b3aae34d71684e4eb56beed82c4