[roll] Roll fuchsia [fidlc] Remove dependency on zircon/assert.h
Add a simple FIDL_CHECK() similar to FX_CHECK(), and use it to replace
ZX_ASSERT_MSG() to avoid dependency on zircon headers (which will be
generated in the future).

FIDL_CHECK() differs from assert() in that it is enabled in all build
types.

Original-Bug: 51001
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/424134
Original-Revision: b1e409eab276353f83b357de3775865250848383
GitOrigin-RevId: a7ae75d9c4ef0d6d79ef1f715848c400a40f2bf3
Change-Id: I51052d7f78a883e2a647b06b2f0ec1c214bc6686