[roll] Roll infra/recipes [jiri] Override HOME for jiri commands
This is for it to generate ssh keys inside the checkout instead of on
the bot's home directory.

Change INFRA_RECIPES=1 to only be set when hooks are run. This happens at two
jiri subcommands:

- jiri update *without* -run-hooks=false
- jiri run-hooks

Stop setting this environment variable the rest of the time, as I think it is
unnecessary.

Include assert to ensure the cwd is set in the context. Update recipe test code
accordingly.

I grepped the jiri sources just to be sure it wasn't looking at INFRA_RECIPES. I
also grepped the jiri sources to ensure the hooks are only run in these two
subcommands.

Original-Bug: 59015
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/424054
Original-Revision: adc887a460b7c5aea565e4f07abc939ff6b7289f
GitOrigin-RevId: db1921ed0013bfc3aa6d2170924cc36f6a490743
Change-Id: I37c685169e17fa7523a6948cc39b82f17111f2c3