[roll] Roll fuchsia [fx] Add options to fx get-build-artifacts
This change introduces three new flags to `fx get-build-artifacts`:
* --expect-one: will fail if zero or more than one artifact is found
* --os: filter by os
* --cpu: filter by archictecture

--expect-one will be useful in the upcoming deprecation of
image_paths.sh, in which we'll look up images and only expect to find
one; the additional filters will be useful when looking up tools of a
specific os+cpu on a remote machine, for use in the -remote scripts.

Also the modes of `flash` and `zbi-tests` are introduced.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/424125
Original-Revision: a9167779cf2307b9770156ae47776e29bb057f91
GitOrigin-RevId: dec9f3b8baa684da1d00e271392a80dbd414fc07
Change-Id: Ic746c2842a608f7e79bdf3b62f90eaa066ddd2bb