[roll] Roll fuchsia [build][alln] Add user.vdso zircon toolchain suite
This CL adds a new Zircon toolchain suite definition for the
"user.vdso" toolchain environment, that will be used to build
the Zircon VDSO with the Fuchsia build.

This only ensures that the toolchain suite definition is correct
and matches the commands from the Zircon build. This doesn't
build anything with it yet.

Original-Bug: 54161
Testing: The toolchain comparison script [1] should succeed

[1] zircon/public/canaries/compare_toolchains.py --clean /tmp/results

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/421577
Original-Revision: 698939183af9829b9646ec85c53c7a88c95887f4
GitOrigin-RevId: 971ce30fad3b7620016a0cf1b2ae350d3f385c20
Change-Id: I35366d591261a50d2c8bef67735648834dd9c7cb