[roll] Roll fuchsia [build]Add fvm fastboot to build
The CL adds fastboot fvm format into the build. The rule for fvm.blk is
kept, but is not enabled by default. It will only be built when the
build is not configured to generate a fastboot fvm format or user runs
"fx netboot" which requires to build an fvm.blk.

Original-Bug: 48803

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/421751
Original-Revision: da0cde5235c5bdfd2a7ba80ec971c56561fe09ca
GitOrigin-RevId: cb5cf90e51c7883ddc488e09dfcff42ff193629a
Change-Id: Ie3dd38c8c31678110245fcfa27c20b854a45c51d