[roll] Roll fuchsia [build] use host_test in go_test
This has the effect of generating a wrapper script for all host-side go
tests. This is a necessary step towards removing support for args in
test_spec.gni, which in turn is necessary to resolve the attached bug.

Also remove a install_host_tools() declaration for size_checker_tests.
size_checker is a host tool, but its tests aren't, so this didn't really
make sense to begin with. But now that the go_test() target doesn't
generate the output file that install_host_tools expected it was causing
errors.

Various other minor clean-ups in go_test.

Original-Bug: 55208
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/420924
Original-Revision: 38fc99ed96a841c16f6118373027339b0dceacde
GitOrigin-RevId: 7889f418b3070aed89139ba61669f8f1880d573d
Change-Id: Icd9f973c96db998ff465ca0f234f28119cff4390