[roll] Roll fuchsia [debugger] Don't recursively enqueue BP steps.
In some error cases we can try to step over the same breakpoint twice.
This can happen if a memory write fails as we try to remove the
breakpoint to step over it (see added comment in debugged_process.h for
more).

This patch prunes the queue so a thread can never be stuck waiting for
itself.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/424657
Original-Revision: 4645a7ec1c54b6c4b52e030a51001d6f31199e1b
GitOrigin-RevId: a30d5a29ad1fffed013b022f503db715438011db
Change-Id: Iea8c767e51f08a7ee86c2bc067da55b13e865854