[roll] Roll fuchsia [build] add fastboot_flash image metadata
Seeds the image rules with fastboot_flash metadata that gets put in the
generated images.json. Nothing uses this metadata yet, but we're working
towards making images.json the source-of-truth for image handling, so
this can be used in the future to replace things like the auto-generated
flash.sh script.

Test: images.json has the fastboot_flash fields (Astro)
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/423724
Original-Revision: 42e8822e9d0ee92d4845b8ce2a4009189313cf90
GitOrigin-RevId: 30d76bd5a9bedd1a66d5f379e990dbafd4157f78
Change-Id: I4fe6f46af4887d03047bd61f980d8d67658814df