[roll] Roll fuchsia [vdso] Address A73 Erattum 858921 in VDSO
Address an erratum specific to Cortex A73 cores where reading the
system timer (what we use for the ticks counter) can produce an
incorrect result during a one cycle window when the lower 32 bits of
the 64 bit counter is rolling over.

For details, refer to erratum 858921 as described in:
https://static.docs.arm.com/epm086451/120/Cortex-A73_MPCore_Software_Developers_Errata_Notice.pdf

This mitigation was already present in the kernel, however we
typically try to read the tick counter entirely in user mode
(bypassing any syscalls) whenever possible.  So, now we basically do
what the kernel does.  If there are _any_ A73 cores on our system, and
we are not forcing zx_ticks_get queries to go through a syscall, then
we install a slightly different version of zx_ticks_get in the VDSO
during VDSO construction time.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/423722
Original-Revision: 9abcb4ce97918c73279656eaed29edd43f83b65a
GitOrigin-RevId: 1c085da4cdba03ca57c6f70061f97c35195cdae7
Change-Id: I7be224b38f15314a4ced5eb49ee159c6ca7fa64a