[roll] Roll infra/recipes [coverage] Use serial log to symbolize and pass in to covargs
Swarming stdout is currently limited to 16MB which results in truncation
and loss of coverage because the symbolizer output is incomplete. Switch
to serial log which should be the same as Swarming stdout but is stored
in a file that's retrieved as task output and thus doesn't have the same
limitation.

Original-Bug: 59087
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/424302
Original-Revision: f0ecdfdc45051de15f7b059f1d5a6e576acafb58
GitOrigin-RevId: c7bbbdc80156cd2f48ea7322c8274975c31923f7
Change-Id: I1023cf434088a10b3c6740a4b620046f2aa398d8