[roll] Roll fuchsia [fidl][rust] Put present first for non-optional clauses
Rust evaluates the first case first. Since presence is
expected in these cases, the outputted code needs to do
less work when it is first.

This doesn't really have much of a performance impact,
but I'm trying to do some minor improvements so that
it is clearer that there aren't outstanding minor
improvements that have an unexpected major effect.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/425196
Original-Revision: 069cc32788f17015cedfbd188cd0959638e39528
GitOrigin-RevId: b3e8fffd8f82fd51ed746c1687a4d105a315f578
Change-Id: Ia4562208b32e53db7a9895d2984c95491b9efeab