[roll] Roll fuchsia [time]Use PushSource API on network_time_service.
This starts to use the time source created earlier, and uses this to add
more comprehesive unittests for maintain_utc (note there is quite a lot
of duplication across these tests - that will get better in the future).

Bug 59408 was a problem with the inspect object falling out of scope
when maintain_utc finishes. This CL fixes that by passing an Arc::Mutex
into the async function so it isn't lost. After diagnostic rework I'll
be using interior mutability and so the mutex won't be required.

Original-Bug: 56857, 59408
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/422887
Original-Revision: 8d03b902c83b680de0532db31f2301dc39eead62
GitOrigin-RevId: 4969a25daa25ef214a26a9b194b828db4bb5f418
Change-Id: I86f7ca1007c3dcd8391829fafe40b0692831013d