[roll] Roll fuchsia [flatland] Have Flatland schedule session updates.
The FrameScheduler has separate APIs for "registering a Present",
which indicates that a client wants to Present in the future and
allocates a PresentId to represent it, and "scheduling a Present",
which tells the FrameScheduler when the client actually wants that
Present to be applied. This change introduces the second API to the
FlatlandPresenter, which must happen after the acquire fences, and
allows the first step to happen immediately.

TEST=fx test -o flatland_unittests
MULTIPLY: flatland_unittests (fuchsia): 30

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/425008
Original-Revision: 9527d3539222863dc725a0af848e403ee95003df
GitOrigin-RevId: 0cbc598182473a6b33be69ecaa56cf42bccb81c9
Change-Id: I2ae188d21efafecb5090fa627e245169fa34c6ab