[roll] Roll fuchsia [bt][rfcomm] Implement initial MuxCommand handling
Introduces parsing support for Frames with MuxCommands. Currently, the
headers of the MuxCommands are parsed and validated. The payload itself
(which is specific to a command) is not parsed. These are marked as
TODOs.

Introduces Encodable,Decodable traits which generalizes the packet
serialization/deserialization process for all MuxCommand frames.

Original-Bug: 58681
Test: Added to bt-rfcomm-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/421194
Original-Revision: fcc35f71dba86537409f155ddcd539e8d987d93f
GitOrigin-RevId: abda6d2eba8357dfb0653b054ba7e9f052a67ace
Change-Id: I6da7adde3297c4231e36b34e1e78b5831768bf84