[roll] Roll infra/recipes [cipd_roller] Support packages across multiple manifests
Storing test duration files in CIPD packages will require pinning the
public and internal test duration packages in separate jiri manifests in
global integration, but ideally the packages would still be rolled
together. To support this, the fuchsia_cipd_roller recipe needs to be
able to roll packages across multiple manifests. To support this, I
merged the `import_in` and `packages` properties into a single
`packages_by_manifest` property that's a mapping from manifest path to
the list of packages pinned in that manifest.

For backwards compatibility, we still support the old `import_in` and
`packages` properties, since deleting them will require a soft
transition with the configs.

I also renamed the "manifest" property to "checkout_manifest" to
differentiate it from the manifest(s) that the packages are imported in.
This will also fix some pylint warnings about loop variables named
"manifest" that clobbered the "manifest" property. This will also
require a soft transition with the configs, which is why I'm including
it in this change – might as well do the two soft transitions at the
same time so I can do everything in 3 CLs instead of 6.

Original-Bug: 46010
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/425434
Original-Revision: e99cca34e6afcf035c049e6ac4f55fbae95ad202
GitOrigin-RevId: 7f1fdcf47bb19536e01a0cb2092d15f6493f4115
Change-Id: I7d25cd324e51d6dd6129cd5cbe5405c0cce9268c