[roll] Roll fuchsia [goldfish] Support composite device in goldfish_control driver.
This change adds support of goldfish-control-2 composite device to
"goldfish-control" driver so that it can access the goldfish address
space device.

Since only one control device is supported on the device (otherwise
there will be multiple goldfish heaps and goldfish display devices),
we do not create a new driver, instead we add some #ifdefs to the
original driver to control which device to bind to.

Currently whether to use the composite device is controlled by a
build-time argument "goldfish_control_use_composite_device" (default
to false) to control which type of device the driver binds to. This
will be switched to true by defualt in a subsequent change.

TEST=goldfish_test with goldfish_control_use_composite_driver = true
     goldfish_test with goldfish_control_use_composite_driver = false

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/424654
Original-Revision: b27a085d64ad6a66cdf3ee97d4845d03f3d9e85d
GitOrigin-RevId: 21b63e16d14fbffcbca300d642519bda76737696
Change-Id: I5b2e839226e760726c5bed4cea274474baf2e557