[roll] Roll fuchsia [fidlgen] Write unformatted generated code when formatting fails
Currently if formatting fails, we output an error but don't write to the
files being generated. This can make it hard to debug fidlgen changes
that result in formatting errors.

This CL changes the behavior so that a file is still written if
formatting fails, but it is written with unformatted text.

Test: success cases are tested through fidlgen tests. There are no tests
for failure cases, but tested manually.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/423917
Original-Revision: f0e8abd6539135388d3fc2b239abd66d91951051
GitOrigin-RevId: 8795afebcfc053ad7dc671ea72bde8fbafb8a1a2
Change-Id: Ie69f979161fb63e08005d0bc3f461a11dd9e7885