[roll] Roll topaz [fidl][dart] Expose unknown constructor for flexible unions
This, in addition to implementing equality on UnknownRawData,
will allow flexible unions to be tested with GIDL.

We discussed whether exposing this constructor is the right approach,
and ultimately decided that:

* Constructing flexible unions with unknown variants is possible, although
in a roundabout way.
* This is useful for testing, not just for GIDL, but also end-developer
testing needs. There is already precedent of that in other bindings
which cannot hide the internal representation, e.g. Rust.
* Usage of this constructor is discouraged in production code.

Test: fx test -v fidl_bindings_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/topaz/+/423174
Original-Revision: 5edf231bffcf737037a588ca8ea124c5a0475a65
GitOrigin-RevId: 4ad3bf97777b1b0b750ea411aceda45ef0ed50a9
Change-Id: Ib060fc9779ce66a21861408b8decbd8a6dfed2b8