[roll] Roll fuchsia [traceutil] Resolve SSH credentials from manifest
The location of the default SSH credentials used in connecting to a
Fuchsia device has changed. This change updates traceutil to use a
manifest to find the appropriate path to the default SSH private key
(identity) and authorized_keys.

The aforementioned manifest is generated by a jiri hook on update and is
not expected to be absent in any Fuchsia tree unless the user has
removed it.

Test: `fx traceutil record` records traces (including WFH setup)

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/419480
Original-Revision: 58f83779a9c3c134c75af290d39077414d905b3f
GitOrigin-RevId: 5e9755a186b89d1104163c2f47228080cef2986a
Change-Id: I49c3dfe7f2dd84d9594696a5c7a0f1e3bb60cdee