[roll] Roll infra/recipes [goma] Allow overriding of GOMA_TMP_DIR from command line
It is sometimes useful to be able to set a custom temporary directory path
for Goma.  This allows one to run a recipe locally.

While running the recipe locally is not a high-fidelity reproduction of
what happens on the build infrastructure, it is quite convenient for fast
iteration when developing a build recipe, and when access to build
infra for testing is inconvenient.

Here is an example command line that starts a recipe, with local goma use.
Note that `/tmp/goma` must be a directory, and must be readable only by
the owner.  The working directory is `/fuchsia/recipes`.

```
env GOMA_TMP_DIR=/tmp/goma \
    GOMA_COMPILER_PROXY_PORT=8094 \
    ./recipes.py run kythe
```

One reason to retarget the temporary directory for goma is if the
default temporary directory is too deeply nested in the filesystem.
Since POSIX limits the path of an IPC socket that goma uses to 105
bytes, if your temp directory is nested in a path that is longer
than 105 bytes (all too easy nowadays), your local recipe will fail
to start goma up.  So, retarget the temp directory and proceed.

A reason to retarget the goma compiler proxy port for the UI is if
you run multiple separate goma instances.  Not that uncommon if you
work on a few projects that all use goma in its different
incarnations.

None of these concerns should be relevant on the actual build
infrastructure, so this change should be transparent when
the code runs there.

Original-Bug: 56412
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/425356
Original-Revision: f59cb0712c5a5e228468f5051542fcfbbe9c13db
GitOrigin-RevId: 59d4294d25255b7a74941f519ab00ca744ef6698
Change-Id: I14293506a28d405a71b491557e7e94be832a6afa