[roll] Roll fuchsia [blobfs] Implement page-in frequency recording
Blobfs will record frequencies of all page-in requests
for all blobs with 8KB page granularity.

These frequencies are stored in the blobfs Inspect VMO.
Since this data is expensive to store (VMO must be >512K),
the size of the Inspect VMO **MUST** be manually adjusted
prior to using this feature.

This feature is gated behind a flag that can be defined
in the BUILD.gn for blobfs.

Subsequent CLs will provide tooling to extract this data
for analysis.

Test: `fx test blobfs-unittest`
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/417934
Original-Revision: d16973f1cdbbae0ab2c54e6e8751c7c91512a554
GitOrigin-RevId: a21c8c60a842858653e887a95376e4a5af54c151
Change-Id: I8aa1572e590fa83c25fd4b541968f7184620deab