[roll] Roll fuchsia [build][alln] Add kernel.phys32 Zircon toolchain suite.
Add BUILD.gn files to support building the kernel.phys32 loader
with the Fuchsia build. This is only used to verify that the build
commmands are identical to the Zircon build ones, thanks to the
compare_toolchains.py script. This doesn't build anything at the
moment.

This introduces a few BUILD.gn that mimic a BUILD.zircon.gn file,
generally in the same directory. These are not symlinks because
only a tiny part of the zircon build file is replicated for this
CL.

Original-Bug: 54160
Testing: RUnnint the script [1] should succeed.

[1] zircon/public/canaries/compare_toolchains.py --clean /tmp/results

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/425315
Original-Revision: f0fd55798d7d7904ccbe2a92f3ca75d08a9bfd07
GitOrigin-RevId: 899b4f463a2f71a8f5a9f11f906d96fa949b5981
Change-Id: I188e31c6e306dc1b14ebadf3df8431c9858c6b24