[roll] Roll fuchsia [fidlc] Parse modifiers in any order
This CL enables fidlc to parse the modifiers "strict", "flexible", and
"resource" in any order. It improves error messages when they are used
on declarations that do not support them, when provided multiple times,
or when conflicting (as in "strict flexible"). These errors are also now
recoverable: parsing continues as if the offending modifier was omitted.

Original-Fixed: 59467
Test: fx test fidl-compiler
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/425897
Original-Revision: 0caaae7673619b1f12ef0dd2828bd4663a327edc
GitOrigin-RevId: 077a2a01ea80d21ebfda8bc441c746407ee2e19c
Change-Id: I5f38091d272247a8807975b7e33245e2f05dcfb2