[roll] Roll fuchsia [rust] Update paste crate to 1.0
rust_icu (which is our own crate) was also updated to 0.3.1, which bumps
its own dependency on paste to 1.0 so that we only need to vendor a
single copy of paste.

ecma402_traits is a "new" crate but is from the same google/rust_icu
repository.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/423703
Original-Revision: 2deb6017c1c12b4981303e03567d02f48f30c096
GitOrigin-RevId: 190591e465e6c32614ee32097a8082386e0cbd70
Change-Id: I9bd4622c8bf3caefdc6c11ce105ac4d612c12279