[roll] Roll fuchsia [debugger] Remove error log for Limbo.
This error writes to the system log whenever limbo isn't enabled which
is the normal state for most developers. It was causing confusion.

The only behavior change is that on error waiting for limbo, the hanging
get for limbo is no longer retried on error.

Minor stylistic updates to the limbo file.

Original-Bug: 58403
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/426017
Original-Revision: 831b93ded06de0feb07ff549e338aab3c8064a1c
GitOrigin-RevId: 712342641f1083a26d41db2a95565dd0140e92ee
Change-Id: Ib583f1f2590ef9bd54472a4ae29b10da6fb562e3