[roll] Roll fuchsia [brcmfmac] Test SSID IE extraction.
fxb/29756 showed a past instance where a scan that sees a beacon that
is missing the SSID element caused a network with SSID of garbage
bytes to appear in the scan list. These tests check that:
- when a scan result's IEs data structure lacks an SSID IE, that data
  structure will be parsed properly--meaning that no SSID IE will get
  extracted into the wlanif_bss_description_t by the brcmfmac driver.
- when called multiple times, the IE value set in the BSS description
  changes as expected.

The new tests increase coverage for the utility method that extracts
these elements into wlanif_bss_description_t.  This utility function
was not the underlying bug behind fxb/29756, but testing the function
explicitly is easy and useful.

Original-Bug: 29756
Test: fx test brcmfmac-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/425212
Original-Revision: bcf758d9a3cf55c22b0caae42dbc9beb3d84002d
GitOrigin-RevId: abb8ae14b10b661e47bfa672101ba112b52296b2
Change-Id: Idf9a8291780b13663e8d334db214efceb4694c2f