[roll] Roll fuchsia [console] Output syslog to serial
This change allows console to act as a LogListener and output the syslog
to the serial console.

This solves the problem of migrating to syslog, but not having access to
logs from the serial console without explicitly running log_listener.

Original-Fixed: 59371
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/423854
Original-Revision: 11a6cd21e8588bf1f78a291752b6bb4f6dfb5772
GitOrigin-RevId: fbe6751bb587f30736a0900e3fdb42d229428f3a
Change-Id: I1e9be3db3fee5893056003a22ec4bacd0316e220