[roll] Roll fuchsia [bt][gap] Refactor LE connection procedure
Refactor LE connection procedure so that pending requests are moved
from LowEnergyConnectionManager::pending_requests_ to their
corresponding LowEnergyConnection in
LowEnergyConnectionManager::connections_ when a connection
completes (before interrogation). When interrogation completes or the
connection is disconnected (whichever is first), notify
the request in the LowEnergyConnection.

This simplifies connection logic and fixes invariants based on the state
of the LowEnergyConnector and LowEnergyConnectionManager::connections_.
These invariant asserts were failing when a connection completed
interrogation while a second connection was connecting or interrogating.

This also fixes a bug where a connection request initiated during the
interrogation of a previous connection could cause the previous
connection's request, still in
LowEnergyConnectionManager::pending_requests_, to be used for the next
connection request. This resulted in a duplicate connection request
that also caused invariants to fail.

When LowEnergyConnectionManager::Connect is called for an existing
connection, call the callback synchronously if interrogation has
completed, or add the callback to the request that will be notified when
interrogation completes. This prevents the possibility of connection
refs being invalidated while the callback is in the dispatch queue,
which prevents needing to intercept the result status.

Original-Fixed: 51186, 52283
Test: in bt-host-unittests
GAP_LowEnergyConnectionManagerTest.DisconnectAfterSecondConnectionRequestInvalidatesRefs
GAP_LowEnergyConnectionManagerTest.DisconnectCompleteEventAfterConnect
GAP_LowEnergyConnectionManagerTest.ConnectAndInterrogateSecondPeerDuringInterrogationOfFirstPeer
GAP_LowEnergyConnectionManagerTest.ConnectSecondPeerDuringInterrogationOfFirstPeer

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/425660
Original-Revision: c9bec9f99f340c097a7a551069256c1c1a5275dd
GitOrigin-RevId: 4608c2e479e5b2f39b30c3a7f5bb4be1e42a0016
Change-Id: Ibc259555074d9a65223bb9a1533ff31bd811f887