[roll] Roll fuchsia [build] Set LLVM_SYMBOLIZER_OPTS in gn_run_binary.sh
This points llvm-symbolizer at the build's own .build-id
directory.  This is necessary to find the unstripped file
for a host tool/test when the stripped file is executed.

For host tools, the variant copying regime makes it slightly to find
the unstripped binary or ensure it's easy to find in
compiled_action().  While that's soluble, for tests, the convention
of finding test data relative to the executable's invocation path
means that running `exe.unstripped/foo` instead of `foo` can break
tests finding their data.  (This isn't a problem in the Zircon build
since it's `foo.debug` rather than a subdirectory, but no Zircon
tests follow that data-finding convention anyway.)  The build ID
solution covers all cases, at least on ELF platforms.  It could
still be worthwhile to arrange to run unstripped binaries for the
macOS case, but the most straightforward way to do that is just to
change the `exe.unstripped` convention to the `.debug` convention,
and there are lots of corners that assume the status quo there.

Also note that some locally-built host tools such as protoc are not
run via gn_run_binary.sh at all, so this doesn't help those cases.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/426595
Original-Revision: ea87992a2a8eb8d2f903debb6fb7eb8bdc3031a8
GitOrigin-RevId: 2e7d0b9a80d558e9b53910416e14f032d6427203
Change-Id: I917d9a8f9e2e176db8bb2bff3642419765955d1d