[roll] Roll fuchsia [elflib] List sources from a given binary
Useful for debuginfod-related efforts, but also for a simple utility I
have in mind for dumping the sources for a given binary (`srcdump`?).
In particular, it would be nice to occasionally use the latter to audit
what's in the kernel or, eventually, physboot.

Original-Bug: 49702
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405939
Original-Revision: d16cc9ba1c2e8efa4b961b07be2601f4e77d813f
GitOrigin-RevId: fa08e023122350291e7eb59384fea01a796c9137
Change-Id: Ieb35ec2ab5431b05b217d9e6c09486a079b4e84a