[roll] Roll third_party/go [fidl][go] Expose unknown handles in unions (1/3)
Ignore unknown data fields in createMarshaler, since the
marshalers created for these fields are unused.

This change avoids the need to create marshalers for these fields,
and also makes it explicit that these fields are not processed as
"data fields" instead of relying on behavior from readOrdinalTag
which happens to allow this to work.

This also prepares for adding the I_unknownHandles field in
fidlgen_go

Test: fx test -v fidl_go_conformance
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/third_party/go/+/426296
Original-Revision: 1a6721874625501aaa08a7643943bded68901109
GitOrigin-RevId: dc779de6a0064a38e794545575dc0d6a898f0721
Change-Id: I481b06bcdc162a04fdc7f4cd16ce37a44c4b354b