[roll] Roll fuchsia [kernel] Mutex: Assert no spinlocks are held when acquiring a Mutex
When attempting to acquire a mutex, a thread should not hold any
spinlocks; if it does, it risks blocking while a spinlock is held,
which is not allowed by the Fuchsia locking rules.

We already have an assertion if we attempt to block while any
spinlocks are held, but that assertion will not catch uncontended
Mutex acquisitions. Add an assert very early in Mutex::Acquire to
catch this case.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/300275
Original-Revision: d833040d382506335f525c03a70b01fd6efe4361
GitOrigin-RevId: 76931cd6b2e63644107393f9fc7a2f3d3e539225
Change-Id: I4f7b6072f390e94d800bb4f98cac1d9a77d3fe3f