[roll] Roll fuchsia [blobfs] Convert blobfs tests to packages.
Remove the "short" integration test. This test triggers some
high-severity logs that result in it being marked as failing despite
passing otherwise. Rather than tracking these down, this test is removed
along the lines of a TODO written by rvargas in the test code that it's
not actually testing something useful and should probably be removed.

Add a flag to the fs-test-support library to control whether the test
is in a package or not. Package tests require an additional path be
injected. The default is packaged since this is the direction we're
going, but the existing minfs and tests for the fs-test-support itself
all required opting-out of this feature to continue running.

Document the "all_source" config which I found confusing when updating
the build.

Original-Bug: 52911

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/421736
Original-Revision: 9bd3b9dd539b280ce3ab8046b4be9a4a3f4a82ab
GitOrigin-RevId: 333c4789822aeb44f9cf0d8d23272ef35eee8268
Change-Id: I3d97dd7554bb46a3ed40a625f319114581d0b075