[roll] Roll fuchsia [build][alln] Build Gigaboot with the Fuchsia build.
This CL allows building the Gigaboot bootloader with
the Fuchsia build, if one adds //zircon/bootloader:bootloader-gn
to the `universe_pacakge_labels` in args.gn

The gernated file will be under out/default/efi_x64/bootx64.efi
(for x64 builds), and is byte-identical to the one built with the
Zircon build (under out/default.zircon/efi-x64-win-clang/bootx64.efi),
except for the RSDS signature, which contains a liberal GUID value that
changes on each linker invocation [1], and the timestamp (which is
built by lld from a hash of the file's content [2]).

This can be verified by dumping their content with:

  objdump -x -D <file>.efi

And comparing the output, which shows a debug directory [3]
in the .rdata section at 0x140015780, where the small section
of distinct bytes appears.

NOTE: A future patch will remove the Zircon build version of the
bootloader, and ensure the Fuchsia-built one is used for core.x64
builds that need it.

[1] http://www.godevtool.com/Other/pdb.htm
[2] https://reviews.llvm.org/D46966
[3] https://docs.microsoft.com/en-us/windows/win32/debug/pe-format#debug-directory-im

Original-Bug: 54924
Testing: See above. Does not affect the build, but produces nearly
         identical binaries as the Zircon build.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/401620
Original-Revision: d5853bb62ba5ee85ab2dc65b7fcfb1f58e2b8794
GitOrigin-RevId: 1d1011868f4e8a33a236159b2af6757d91bf524f
Change-Id: I993a9beea5fae18451d6767c09e75a403bdd46ab