[roll] Roll fuchsia [components] Speculative fix for Hub v2
The stop event could be fired after destruction which would cause the
hub to crash component manager. This CL ensures that the stop event
happens before destruction.

This also fixes an underlying shutdown bug: stop and shutdown must be
serialized with respect to one another so that we cannot stop and
restart a component during shutdown. The shutdown bit must be set to
indicate that a component has stopped and cannot be started again. Prior
to this change, it is possible for that component to not be set during
shutdown due to a race with a stop.

This CL makes Stop an Action and adds a special case to make sure that
stop and shutdown actions serialize with one another.

Original-Bug: 58378
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/425654
Original-Revision: c3b4488097e2f0721248f4bfd743e71fb9a440db
GitOrigin-RevId: b681c00fd324096fcf6d612a2676e0975a09df7c
Change-Id: Ib1d5e978da57ad422f641b19a72d7c0fdc170974