[roll] Roll fuchsia [fuchsia-hyper] Fix panics, add new errors
In http://fxbug.dev/59391, the netstack crashed due to minfs filling up.
This caused the package resolver to panic when it tried to communicate
with fuchsia.posix.socket.Provider in order to resolve an interface name
to a scope id.

This patch switches from panicking to returning an error when it cannot
communicate with fuchsia.posix.socket.Provider. In addition, this changes
fuchsia-hyper to error out if the interface index is larger than 32
bits.

Test: Added tests for these cases.
Original-Fixed: 59384
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/423695
Original-Revision: 1ebe6012c7b28f281d7a618730495e0974eb3908
GitOrigin-RevId: 7102b568ff31040da8ca20549ac0f5fe3bd6521f
Change-Id: I3a5c607de394d199f2ba59af010db3bbeabd17a6