[roll] Roll fuchsia [kernel][vm] Test for merging markers
Adds a unittest to validate that the migrate_fn callback of
VmPageList::MergeFrom is not called for markers. This is the current
behavior, but in the implementation and interface description, but it
is easy to violate when modifying the page list implementation and
there are no other tests for it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/425556
Original-Revision: 7e9fbf1bfd3221586a9e59f2a36e8e2362f1cde8
GitOrigin-RevId: 7a90844844d6a9138b87d5df471b52d3cdf00173
Change-Id: I37e6467888546e266b2ec50a53814e1bd82f7c63