[roll] Roll cobalt [Cobalt 1.1] Enforce event_code_buffer_max
Limits the number of event vectors which are logged
for a single EventCodeAggregate to at most the
event_code_buffer_max of the corresponding
MetricDefinition. In addition, limits the number of
event vectors which are included in an Observation
to event_code_buffer_max.

Event vectors are prioritized in order of first
arrival.

Original-Bug: 58854
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt/+/420574
Original-Revision: b7c4423540b17bfd3773d9e3eaeee112ce6b3098
GitOrigin-RevId: 290513dec9360017dc99c15e63878a80ebd6eaa9
Change-Id: I4f38128c9fb314f8013a89298b696293f62aeb84