[roll] Roll fuchsia [gidl][rust] Use both handles from syscalls that return pairs
This CL makes GIDL conformance tests and benchmarks for Rust behave the
same as Go and Dart with respect to handle creation. Before, for handles
that come in pairs (channels), it would return one endpoint and discard
the other. It now uses a factory object that reuses the "extra" handle.

Test: fx test //src/lib/fidl/rust
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/426897
Original-Revision: bc536e0d86c37bfee000425772044fe21716575b
GitOrigin-RevId: 06a0c5ebb05ef533c40919c03358452f688212ae
Change-Id: I48af7935d9b91aee9d16b3b9d03ba9cb13ebb94e