[roll] Roll fuchsia [dart][build] Dart test deps use dart toolchain
Rewrite all dart_test deps to use the dart_toolchain. The symptom of
this is that a FIDL dependency will be built against the toolchain of
the test, which is the host toolchain if this is a host test. And that
fails.

There's already a non_dart_deps parameter for specifying non dart deps.

Fix: 59254
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/426975
Original-Revision: 72ba74d05b8002dd9586adcdf0d4510aa5aa94c6
GitOrigin-RevId: 758b783581e14f0a6bebb112910bae9021453903
Change-Id: I96bd419ae46aa92ee3e63d16dacc6cf568540d2e