[roll] Roll fuchsia [docs][tests] Split the "Running tests as components" into concepts and how-to guides.
This change only splits the existing "Running tests as components" guide
into a concept doc and a how-to guide.

Further editing the content of these docs will take place once this
change is submitted -- this allows the updates of content to be viewed,
compared and commented on Gerrit.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/426023
Original-Revision: 797f1503d56ddf8d8dc6ec2c50ef0841d7ec4401
GitOrigin-RevId: 630540e2742e783352439d3a53452cebce5d7e42
Change-Id: I821fbe45f5bf4becff451ce43f41de2db8338055