[roll] Roll fuchsia [a11y] Implements recover A11Y Focus action.
This is a simple action to be executed when a semantic tree is updated. The A11Y focus will be checked to see if it still exists. If not, it is cleared, and as a result, so is the highlight.

This action is expected to be improved as we learn what are good ways of recovering the focus.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427015
Original-Revision: dde7b0edd42d8315fc85967881e7fcb1ba5a3bc4
GitOrigin-RevId: 0d75bb73aa4b1a70d64ee59cf94d532ed57c55be
Change-Id: I1478d43e5084bc61b368ac54024477f058c14a33