[roll] Roll fuchsia [zxcrypt] Allocate slices from 0, not 1
When running atop FVM, zxcrypt needs to ensure that it has enough space
allocated for both its own superblock and backup superblock, but also
it needs to ensure that the zeroth block of the FVM that it exposes is
also preallocated, since that's the usual behavior of the FVM itself.

An error in calculation caused us to preallocate an additional slice
here which zxcrypt never used and which the volume sitting atop it might
also never use.

Tests: added new test verifying that when presented with a new volume
(0th slice allocated, all others unallocated), we allocate exactly one
additional slice in `Init()`, and that if we already have two slices
allocated, we do not attempt to allocate additional slices.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427257
Original-Revision: eaec7f2310e85c8f5f05655a0b7331126647537f
GitOrigin-RevId: ca27d14f656cb2268b8d850fdff901ca659156e2
Change-Id: Ief45afee0fa486bf7c73fec89effea3865e6c950