[roll] Roll fuchsia [build] disallow fuchsia tests to set args in test spec
The component manifest already has a mechanism for specifying args which
get passed to the component itself.
The args feature was only being used by prebuilt_test_manifest to change
the runner which runs the component, but we don't want to allow that,
since we rely on the runners to enforce very important policies on
tests (e.g. timeouts, log severities).
This change closes a loop-hole that would allow tests to subvert these
policies, and it also simplifies the build system.

A later change will remove support for the custom command from the
testrunner.

Original-Bug: 58561
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/418483
Original-Revision: b247738bc374e45d316695b56da1475e01800d1b
GitOrigin-RevId: d3fcdb720c29c269cfa8b581b402f382071dc6df
Change-Id: I0f3a6ed263c4334385916cdb78803141e41a5d9e