[roll] Roll fuchsia [ext4] Fix unsafe construct
We can't just `Arc::from_raw` on a `Box::into_raw` as this will expect
the pointer to be part of an `ArcInner` and as such improperly use heap
memory located prior to the pointer as part of this struture, notably
the Scudo chunk header.

This was noticed when trying to update Scudo (fxr/343518) and lead to
various errors.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427095
Original-Revision: 30888486d536e7293ddd791dc91818b37034e4d2
GitOrigin-RevId: 443e3924769e69fe5fde635ff246709428c6cd36
Change-Id: Ib910b1334b8f7252ee7f0bda7a307f1ac877d45c