[roll] Roll fuchsia [zbitl] Don't assume test payloads are NUL-free
This is a brittle assumption, which will be invalidated once we start
looking at expected ranges of storage across multiple items. In this
change, we alias std::string so as not to convey that we are dealing in
text - and replace all usage of EXPECT_STR_EQ with EXPECT_BYTES_EQ.

Also, we sneak in a standardization to make use of `storage_type` and
`payload_type` in the test trait signatures.

Original-Bug: 57764
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427355
Original-Revision: 995b62ee8458127cc4570b425da8ed2cbea2b4e6
GitOrigin-RevId: 8d4ecd639f01bcf0bc68994a90e73c995782ecc4
Change-Id: I4d7d98c964afa14f5d6ca0cd7c93711a90bc39c1