[roll] Roll fuchsia [wlan][policy] matching mixed networks
This change allows us to auto connect to mixed networks if the network
is saved as either of the security types in the mixed security type. Ie
if we save a WPA3 (or WPA2) network and see it as a WPA2/WPA3 network we
will auto connect to it.

Multiply: wlancfg-tests
Original-Bug: 53834
Test: fx test wlancfg-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/424162
Original-Revision: 969ca2e07c68ae3552eea81045f95e4424a61b88
GitOrigin-RevId: 4026338bb70efb3f28aa3ff0b91cd5b74e93627d
Change-Id: Ia2a8b7fd0254828a155d0d80f32e781116f55940