[roll] Roll fuchsia [audio_core] Device clock is clone of driver clock
In preparation of synthesizing the hardware clock from position
notifications, AudioDriver maintains a separate AudioClock that it
will rate-adjust as part of the clock recovery process. The
existing AudioClock (passed upward to the output or input mix) is
now a read-only clone of the AudioDriver's underlying zx::clock.

We only create this separate clock, and perform hardware clock
recovery, if the device is NOT in the Monotonic clock domain.
For this reason, AudioDriver must wait until it has retrieved the
clock domain from the actual audio driver, before creating the
AudioClock.

Original-Bug:  46631
Test: AudioDriver/AudioDevice tests reworked to accommodate

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427151
Original-Revision: 81980bec8873aa0645d8a47e3232e72b14067277
GitOrigin-RevId: 98c8eee5607ba5178dc195202a1de04c3c915849
Change-Id: I3183cfbfbe4a08f3d6c0dcd361c0449f9a673bbd