[roll] Roll fuchsia [testrunner] Reconnect if failed to run snapshot.
We only try to reconnect to SSH in between tests, so if we lose the
connection after the last test and before we run the snapshot, it will
fail with "ssh is diconnected". We should try to reconnect and retry
running the snapshot as well.

Original-Bug: 60039
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427267
Original-Revision: 52ace21c203ad4b4a4fff45e4f99ac60cf72d0c6
GitOrigin-RevId: 9fe7aa24d6ac67366a953f1c666be2c901865758
Change-Id: I4120901d9825d8ab01fc984ed0159e9063760730