[roll] Roll fuchsia [build] Fix LTO build.
A previous CL moved the build of the Gigaboot bootloader to the
Fuchsia build. This binary in a Win32 PE+ file that needs to be
linked with the "lld-link" linker, which expects slightly different
command-line parameter names to control lto parallelism and caching.

This broke the LTO build, which is not on CQ, hence why it went
unnotticed at the time of submit. This CL fixes the issue by ensuring
the right parameters are used for this specific case.

Original-Fixed: 60060
Testing: Fix the LTO build (e.g. --variant=thintlo)
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427220
Original-Revision: f4da3522c5eb2a3d7ed9d267cda5626a7f9f2aa8
GitOrigin-RevId: 6663682f8dca0e9c86e74fa34a701656f0dcd451
Change-Id: I905ec8839078cfbb1bd1a76bd3360cb198828241