[roll] Roll fuchsia [fuchsia-hyper] Fix panics, add new errors
In http://fxbug.dev/59391, the netstack crashed due to minfs filling up.
This caused the package resolver to panic when it tried to communicate
with fuchsia.posix.socket.Provider in order to resolve an interface name
to a scope id.

This patch switches from panicking to returning an error when it cannot
communicate with fuchsia.posix.socket.Provider. In addition, this changes
fuchsia-hyper to error out if the interface index is larger than 32
bits.

This is a re-land I47224872f2ba803e62051c734126b98e891eb317, which was
flaky because the `test_parse_ipv6_addr_handles_large_interface_indices`
was making a synchronous call to fuchsia.posix.socket.Provider on the
main thread, but it's unsupported to make a blocking call with a single
threaded executor.

Multiply: fuchsia-hyper-tests

Test: Added tests for these cases.
Original-Fixed: 59384

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427626
Original-Revision: 731b4abb30908eccbe9beca0470f145afd936e8c
GitOrigin-RevId: 212d519d263221ea209d626b35b3362267bf36ac
Change-Id: Ic5e36111cda8f1e3f0a2f5833bd2f436569a2d71