[roll] Roll fuchsia [usb] Fix invalid assumption about port ordering
Per USB 2.0 specification revision 2.0 section 9.1.2,
enumeration order on a hub is not guaranteed to be sequential
in order of port number. This was accurately reflected in my hub simulation,
but the test didn't account for this behavior. This CL updates the enumeration
test to allow for enumeration to occur in any order.

Test: fx test "//src/devices/usb/drivers/usb-hub-rewrite:usb-hub-package_test_usb-hub-package_component(//build/toolchain/fuchsia:x64)"
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427463
Original-Revision: 20e4ac46ed581e8dc6837daa4e554c333b55dce4
GitOrigin-RevId: fc06840112519a7bce9999eb655c475e3e6b3c3b
Change-Id: I39b15be0fa843235dbb5a0fd2df39a01b43119e3