[roll] Roll fuchsia [build] Add documentation for gen-latest-commit-date args.
fxb/58677 is pointing to the fact that the formats for our backstop
files are not well documented and will be even less well documented once
we remove the kernel clock and its command line argument.

This CL adds a little more documentation by describing the expected
output in the script that generates the files.

Original-Bug: 58677
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/426015
Original-Revision: c537e317eefb8436781e73d62d8d28389df2e829
GitOrigin-RevId: 9defd82d03ab8ac2f0fef3e04b890f8f4b90fe88
Change-Id: I798ca5716469aeff62e78fd8529924e880c8ca17