[roll] Roll fuchsia [build] Set *_SYMBOLIZER_PATH in environment of host_tool_action() runs
The Zircon build always had machinery intended for this, but it
hadn't been fully wired up.  Note that the Zircon build already
runs the unstripped binaries for host_tool_action() though it
heretofore failed to point to the symbolizer so it would only get
full symbolization if llvm-symbolizer or addr2line that works is
in the user's PATH already.  The Fuchsia build already takes care
to set the environment variables via gn_run_binary.sh, but it
runs the stripped binaries so available symbolization is minimal.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/426564
Original-Revision: d6c2fd3513b35f01c030900cc686777c256742bd
GitOrigin-RevId: 8c1aff37932a9e4384a6fd29d66e4b912b109fee
Change-Id: Ic4e2cb44e584b781d1950d8304e12eee5237cd3c