[roll] Roll fuchsia [build] Allow metadata for shared_library() instances.
Building the Musl C library with the Fuchsia build requires passing
metadata to a shared_library() instance. This was not possible, and
this CL fixes the issue.

For the record, the metadata is used to populate the sysroot.

Original-Bug: 54161
Testing: Should not change the build.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427856
Original-Revision: 4686e8e94f60704917dc200fbab55409100ebf66
GitOrigin-RevId: db7b52666ff89eebd19de7738c9428439a273afb
Change-Id: Ia0bab70213d158faed1774571aa99f1798dab9a7