[roll] Roll fuchsia [flatland] Flatland unittests cleanup part 1.
Many of the Flatland unittests are actually closer to integration
tests. Specifically, the ProcessMainLoop() function replicates the
main loop that now lives in the flatland::Engine, so any test that
relies on that function more-than-likely belongs in the Engine
unittests.

The goal is for the Flatland unittests to strictly test the API: what
valid inputs are, and the outputs (in the form of UberStruct data)
that it should produce.

This change contains smaller fixes to prepare for deleting the
ProcessMainLoop() function:
1. Removes ProcessMainLoop() from every test except for the Link
   tests. All other tests that depended on this function now directly
   read from UberStructs.
2. Add a function to Flatland for getting the TransformHandle for a
   particular ContentId to verify presence in the local_topology.
3. Remove unnecessary RunLoopUntilIdle() calls.

Part 2 will include cleanup of LinkSystem::UpdateLinks() and the
removal of ProcessMainLoop().

TEST=fx test -o flatland_unittests
MULTIPLY: flatland_unittests (fuchsia): 30

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427621
Original-Revision: 793607469e936e3bf4c81f5c56ff3d07304604f5
GitOrigin-RevId: 4e67f8f68dd4f162c5bc5d7b79799b51a558e3a7
Change-Id: If96bdef26ec7c6559832f7e79be24c16eee4e38d