[roll] Roll fuchsia [sherlock] Add clock ID for h264 video encoder clock gate
The h264 encoder shares a clock register with the decoder.
Add clock ID and bind to it so we can synchronize it's use.

As the driver that binds the composite device is in a different repo,
the test for this is in a follow up CL. Existing device enumeration
tests ensure this doesn't break anything.

Test: fx test h264_encoder_test
Original-Bug: 45193
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/428056
Original-Revision: 076d785e2af85b5a55f1e04cdd3c154eda959880
GitOrigin-RevId: 9686d8619df6ce4b834827a4d544c42892b934ba
Change-Id: I5d25293568e131151329b1d961207bf7b50b8fd5