[roll] Roll fuchsia [goldfish] Support VkBuffer allocation on goldfish HostVisibleHeap.
Now goldfish HostVisibleHeap can allocate VMOs and set up
renderControl Buffers for VkBuffer usage.

Currently whether it is a Buffer or it's an Image is determined by
the constraints: If the constraints contains ImageFormatConstraints,
then we regard it as a Buffer, otherwise we regard it as an Image.

TEST=add tests testing buffer allocation to goldfish_test
Original-Bug: 59805

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/426320
Original-Revision: d501398f634b844044e3182bc7e4b95f7097c616
GitOrigin-RevId: 305543a217782ff75be7b89ba43498857d7f10da
Change-Id: Iaafa65c8545cef2eef33adcf8e41e212b1d71810