[roll] Roll fuchsia [weave] Reduce the max sleep time for select to 1 second
Currently the max sleep time for select is set to 10 seconds.
This causes the runloop to sleep longer and timer/task processing
is delayed. Reduce it to 1 second. And also wakeup the
runloop after posting a task, so that the task runs immediately.

Original-Bug: 59242
Test: Manually verified that the runloop wakes up within 1 second
      and whenever a task is posted.
Test: fx test weavestack-unittests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/423038
Original-Revision: 24af71ce7e3de9f310416962cc0401e7ff45d69d
GitOrigin-RevId: a516cc0c7af60ee6926ce2ff0c8d8b0573639718
Change-Id: I4a6c6e13b88d737067a64ce3d3cdc1fd1828c3c3