[roll] Roll fuchsia [codesize] Simple test for fx codesize
This tests running "fx codesize --help" exits successfully
with code 0.

Currently, this cannot run on CQ because a dart binary
isn't available in that environment, but plans for
support for the runtime in CQ are in progress and
it can be run with fx self-test in the meantime.

Test: fx self-test fx_codesize_test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427017
Original-Revision: f37304b338f8e36c73cc63f8209527416e4ef29d
GitOrigin-RevId: c3abe28fc22329eb35d1fbca374ebfb4d9e42731
Change-Id: Ibcb13ed9e2163a893ab32b2d03dc766752725ee0