[roll] Roll fuchsia [wlan][brcmfmac] Fix brcmfmac firmware event debug logging
This change makes two modifications to debug logging. The first is
that firmware event names, when logged, will no longer print as
"nodebug" but instead with the firmware event name. Firmware events
are generally only logged in errors or when debug logging is enabled,
so this has no impact on production performance.

The second change is to fix an error in brcmf_fil_get_errstr(). The
bcme_status_t, when transformed to a string, now prints its own name
and instead of always printing BCME_STATUS.

Test: None necessary. Logging changes only.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427928
Original-Revision: 9baab964aee53585a71df9f087e667b12addfa10
GitOrigin-RevId: 2fc1aed2bdeba28601671fbbdfdc1b0a2971c4cb
Change-Id: Iceaeee51596c3d6cbff4f469ebbc7e407e06a976