[roll] Roll fuchsia [license] Support parsing of different authors
Currently we have to add a golden pattern for each license type
and the author. This results in 200+ patterns and incorrect showing
of the matches. This CL adds a map for authors and matches.

Original-Bug: b/167403181
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/426974
Original-Revision: b89c2f6e5f9b80220e0a421c3b9303c0f32a1713
GitOrigin-RevId: 8fe3e7af8e3a58a49cae24ffecde1dea990a2224
Change-Id: Iee61ad342c5a9cfb1ab619c7537ddb0ea9f7787f