[roll] Roll infra/recipes [fuchsia_cipd_roller] Check tags for all platforms
Previously, the CIPD roller recipe attempted to roll all the specified
packages, across all platforms, to the same tag. However, when searching
for a tag to roll to, the roller only checked the the current platform's
version of each package because the CIPD CLI automatically substitutes
"${platform}" in any package names with the name of the current
platform.

For example, if a package had versions for linux-amd64 and mac-amd64,
the roller would only check the tags for the linux-amd64 version of the
package (because the rollers always run on Linux). If the mac-amd64
version was out of date and didn't have any instance with any of the
tags that the latest instance of the linux-amd64 version had, the roller
would still try to roll to one of the tags found on the linux-amd64
version. It would then fail to run `jiri resolve`, which tries to look
up package instances by tag across all platforms, because the chosen tag
didn't exist on any instance of the Mac-specific package. For example:
http://ci.chromium.org/b/8868906404994183200

We can fix this by using `jiri manifest` to get the platform-specific
versions of each package that are included in the checkout, and then
search for tags using the platform-specific packages instead of the
packages with "${platform}" in the name.

With the fix (re-running the failed build from above):
https://ci.chromium.org/swarming/task/4eb5254f3c438110?server=chrome-swarming.appspot.com

This also let me remove some of the convoluted logic that parsed the
stdout of `jiri resolve` to find the names of platform-specific packages
in order to present them in the commit message. That's no longer
necessary now that we use `jiri manifest` to find the platform-specific
package names.

Original-Bug: 60085
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/428354
Original-Revision: 0c99421e93d489470440335106c63ce31eaf7d26
GitOrigin-RevId: 8051a83a5dd8e09c9c3746a3b56faf0775d205cf
Change-Id: If81e2c4beccaa9584cf21f1f2898e09e871328b9