[roll] Roll infra/recipes [led] Pin led as temp fix.
This is a temp fix until crbug.com/1073058 gets fixed. Currently, the
led module downloads the latest version of led from CIPD and uses it.
The problem is that there could be (and has been) changes upstream in
led that break our recipes. Thus, we want to have a pinned version of
led so that upstream changes won't break us.

This change creates a wrapper module to inherit from the recipe_engine's
led module and overwrite the _ensure_led() method to use the pinned
version.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/427734
Original-Revision: ab9ce9f653c66de0518f283a88df11226d307fd1
GitOrigin-RevId: d79b7c496c07ec15e064a5036b28dab0b2b1fe4e
Change-Id: Id0142f6c1fcdb5935d00071716c1a7a752f4b6ac