[roll] Roll fuchsia [tools][fidlc] Improve parsing of response files.
This CL removes an important restriction from fidlc, i.e. that
a response file, if used on the command-line, is the only argument
allowed when invoking the program.

Instead, the response file is just read and inserted into the
current command-line, allowing the caller to provide other arguments
before or after the response file, or even using several ones in
a single call. Response files still do not nest though.

This CL is necessary to simplify build unification. For more
context, see the comments on this Gerrit CL:

  https://fuchsia-review.googlesource.com/c/fuchsia/+/427858

Original-Bug: 54161
Testing: Should not change the build.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/428074
Original-Revision: e7cd43ab59f05edc79ff2302536fa70c6f47e888
GitOrigin-RevId: 93592a9289f94800d28c506c6e324462698f9640
Change-Id: Ifb77ec81a90d5fefece05dd1bdf0b4bfbe8d89c3