[roll] Roll fuchsia [fx] Use get-build-artifacts for pave-remote; extend to `netboot` and `pave-remote`
Right now, we rely on build-archive.tar having everything needed to
pave; that gets rsync'ed and unpacked. This change updates the command
to instead use fx get-build-artifacts (which uses `jq` on images.json)
to find precisely the set of images needed.

This change is beneficial for four reasons:
(1) it is more efficient, only building and rsyncing precisely what is
needed - and avoiding unpacking
(2) it readily extends to the similar use-cases of remote netbooting and
running a ZBI test (see follow-up CLs)
(3) it removes a dependency on the build archive (which is in the
process of being deleted/deprecated).
(4) in the cases where (host OS, host arch) == (remote OS, remote arch),
we now no longer need to build bootserver locally.

We introduce `fx netboot-remote` and `fx pave-zedboot-remote`
along the same lines.

Test: tested locally on my glinux laptop - and enlisted someone with a
mac to test locally as well.

Original-Fixed: 58096
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416964
Original-Revision: 7377ba6739278af7fa22c65f594fc0b20e51457d
GitOrigin-RevId: 7fd44e75f0040615e01d595fb174c98128b3f163
Change-Id: I5005f03647e6afddc1d642c4a3d735a7bc5f6055