[roll] Roll fuchsia [SetUI] Introduce Event to Handler MessageHub
This changelist introduces the Event enum to the Handler
MessageHub payload and defines the Changed event
type there. For this event, the SettingType has been
dropped as it is redundant for the SettingProxy.

This change is in anticipation for an upcoming addition of
new event types.

Original-Bug: 58181
Test: fx test -o setui_service_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/426314
Original-Revision: 952ee6ca642d74ab70e258320289de26eb8b4237
GitOrigin-RevId: 547c8372dd2099cbfb7a4bfd29059b1521aa5cb1
Change-Id: I1620415f6e8a0dbc34ab5ed14ee49f0251e38195