[roll] Roll fuchsia [build] Ignore variant_target in toolchain_util_action()
This fixes the definition of toolchain_util_action() when
it is called from the Fuchsia build. When migrating the
arm64 boot shims, the `variant_target` parameter will be
set.

Instead of asserting, just ignore the parameter because
it is not used in practice: i.e. this only makes sense
when the kernel is built with a different variant than
the boot shims, something that is not supported anymore.

Original-Bug: 59918
Testing: Should not change the build.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/428538
Original-Revision: bc519d169fb26d05f43fde53d759adc2cef4e198
GitOrigin-RevId: 660ca670b30637e451ec3ade2cf40d0212a41c7c
Change-Id: Ieedbadeb951bc86265783f06e2207a4dbee0700b