[roll] Roll fuchsia [build][alln] Prepare zircon script for host tool migration.
Host tools like lz4 and zbi are soon going to be moved entirely
to the Fuchia build. This CL modifies a Zircon script that is
invoked from the Fuchsia build, but expects some files to
be located in the Zircon output directory.

This simply replaces all occurences of BUILD_DIR with
ZIRCON_BUILD_DIR, and adds a new variable ROOT_BUILD_DIR
to point to the Fuchsia output directory instead.

The latter can be set manually with -R <build-path>, or
is simply guessed from Zircon build dir otherwise.

As more items are moved from the Zircon to the Fuchsia builds,
lines of this script will be updated progressively.

+ Remove obsolete source file that is no longer compiled
  by anything, and the script lines that reference its binary.

Original-Bug: 49742
Testing: Should not change the build.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/428536
Original-Revision: 92d6fdcd25170704427502a2fa47b94591956a68
GitOrigin-RevId: 50776e7e0f4026f733ead66f6c3a66df43a824b4
Change-Id: I49f9b6728cf0f44b3651d1ac94ef0db051beb3bf