[roll] Roll fuchsia [build][alln] Restore obsolete legacy libraries entries in .gitignore
This restores a few entries in zircon/public/lib/.gitignore for a few
libraries that were already migrated to the Fuchsia build. The reason
to do this is to avoid lingering untracked directories to appear
in 'git status', because the corresponding files are still in the
workspace (they were not managed by git but auto-generated).

It doesn't change the build in any way, it just makes it easier
to avoid polluting workspaces with untracked files during build
unification.

A comment is also added to the file explaining why this is needed
and how to manage this file during future legacy target migrations.

Original-Bug: 58463
Testing: No change to the build.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/428535
Original-Revision: 2c50f189107f9a1e553ed78fc6bce6b78434373d
GitOrigin-RevId: 63b6c503bbca4d74c81dccedfc2a0f0c561109b1
Change-Id: I003c7e504e905bbeaaad91196fffc3978141243d