[roll] Roll infra/recipes [fuchsia_roller] Fix urlparse bug when rolling multiple commits
When given a plain hostname without a scheme, urlparse.urlparse()
considers the entire input to be the URL path:

  >>> urlparse.urlparse('foo.com')
  ParseResult(scheme='', netloc='', path='foo.com', params='', query='', fragment='')

This caused issues in the write_commit_message() function, in which we
run `origin_host = urlparse(origin_host).netloc` for every commit in the
roll. If there are multiple commits, `origin_host` will become empty
after the first commit because the following always returns an empty
string:

  urlparse(urlparse(<anything>).netloc).netloc

That causes `api.gerrit.change_details()` to fail (unless the build was
triggered by a gitiles commit, in which case the Gerrit module will fall
back to the Gerrit host for the triggering commit). For example:
https://ci.chromium.org/b/8868852257441362480

I reproduced this failure in a test case in patchset #1:
https://fuchsia-review.googlesource.com/c/infra/recipes/+/428695/1/recipes/fuchsia_roller.expected/roll_from_non_https_remote.json#416
We didn't previously have any test cases for this recipe where the jiri
manifest contained a gerrit host for the rolled repo *and* we were
rolling multiple commits, which is why it was possible for this bug to
land.

Original-Bug: 54464
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/428695
Original-Revision: c61b533674c9d82297418d3ad8cb881e20cf05aa
GitOrigin-RevId: f66509ffdd921fd3d5a8af5f30c60c683e993635
Change-Id: I1deb771ed32c99e2489024a5a4b4f34afbfa9694