[roll] Roll fuchsia [event-queue] Add try_flush() API
This change adds a try_flush method to event-queue's control handle that
produces a future that completes when all currently queued events are
dispatched and acknowledged by actively connected clients.

Normally, users of event-queue want to fire off events to any listeners
and then forget about them, but if a service emitting events wants to
stop or restart, for example, it needs a mechanism to wait for all
enqueued events to be sent to listeners before it can do so.

Note that, normally, it is by-design that listeners of events can't stop
the service from making forward progress.  To ensure that a misbehaving
listener can't block a service from stopping, this API will timeout
after a configurable duration.

Original-Fixed: 60146
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427264
Original-Revision: 9a4bf8abca91fb3145ef27fda087e9016e85b0ef
GitOrigin-RevId: 599a2d9c8ccc9c34862d45e35d92d85adc9d302f
Change-Id: I2da45b0cca15a53172528b86dd46e253f90e3a6b