[roll] Roll fuchsia [fdt] Always unzip the prebuilt to run
Before we were opening an already unzipped prebuilt binary if it
existed. The prebuilt system is only updating the .zip file so an
already unzipped binary might be stale so this is not a safe check.
Instead we always do the unzip step so that we actually pick up updates.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/428317
Original-Revision: 547535c7f2c1b2996c1f6afcffeca4b31910189a
GitOrigin-RevId: e3f12b1c71ea93e981f7ef8a41644862f8f6a23d
Change-Id: Ie488e0310d1877599e49e06984f6833bf927c905