[roll] Roll fuchsia [llcpp] Separate client channel/binding ownership.
This CL aims to solve the following problem:
Presently, the lifetime of the channel is a subset of the lifetime of
the AsyncBinding, thus a std::shared_ptr<AsyncBinding> is used to
guarantee safe access. The unbind operation, executing on either a
dispatcher thread or a thread shutting down the dispatcher, blocks on
the release of all references to the AsyncBinding. For a server, all
additional references are transient. For a client, this could block on a
synchronous call to another process, potentially locking up the entire
process.

This CL solves the problem by separating the channel and AsyncBinding
lifetimes and sharing ownership of it between the AsyncBinding and the
ClientBase. This manifests as a new ChannelSyncRef class. The ClientBase
and AsyncBinding (via the OnUnboundFn) each hold a
std::shared_ptr<ChannelSyncRef>. Temporary references are generated
from a std::weak_ptr<ChannelSyncRef> for the duration of any outgoing
calls. On destruction, the ChannelSyncRef can be configured to signal a
sync_completion_t, enabling a new fidl::Client API WaitForChannel()
which is used instead of the OnClientUnboundFn to recover the channel
safely. A consequence of the new API is that the channel can no longer
be recovered via the OnClientUnboundFn, thus zx::channel argument has
been removed.

NOTE: Another possibility is to simply just post a task to invoke the
OnUnboundFn from ~AsyncBinding() rather than blocking in a dispatcher
thread. However, this is problematic in the case of dispatcher shutdown,
as the OnUnboundFn would then need to be executed on the thread on which
~AsyncBinding() is running. This could be nested within other user-code,
potentially triggering a deadlock.

Original-Bug: None.
Test: fidl-walker-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427244
Original-Revision: 7323de79cf3995235d7ae91a398b3cf1495b72c9
GitOrigin-RevId: 51a703ce2ed2e81b0c53b3966596c3ebf4047645
Change-Id: I98a29238db9018e67fc86901fe92fcf1c5b7c248