[roll] Roll fuchsia [wlan][sme] Make ConnectionLoss event accurate
Currently the ConnectionLoss event is triggered irrespective
of whether it was the AP or STA that initiated the disassoc/deauth.
This in turn results in inaccurate metrics.

This change ensures the event is triggered only when the disassoc/deauth
is triggered by the STA.

Bugs: 37823
Test: New tests added to wlan-sme-tests
Multiply: wlan-sme-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/428320
Original-Revision: ba91288c552003152762ebd2cabbc488cfd85b40
GitOrigin-RevId: 413ad4d2e72fa30274baff60c515140ea0062986
Change-Id: I0bc57e8e61dd71a53ee917a87bc41e433c7f3f13