[roll] Roll fuchsia [fxtest] Exit `fx test` if build fails or cannot connect to device
When I'm testing a stack of CLs, I'll frequently run this loop:

```
% while fx test ... && git rebase --continue; do sleep 1; done
```

However, it turns out `fx test ...` only exits with a non-zero status if
the test fails, but exits with 0 if the build fails, or the test cannot
talk to the device. This patch changes `fx test` to fail in that case.

Test: Added test to make sure `fx test` will exit with a non-zero status
if the build fails or no running package server.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427627
Original-Revision: 749ca63d25ea54882349dbd2e1fe136e184030cb
GitOrigin-RevId: 92f713daf7d4746b4bb03cb604ea7e88c069686a
Change-Id: I125dd49fcc0df21f7d0145dbad6a564a60c0feaa