[roll] Roll fuchsia [wlan] Cleanup prototypes of wlan final methods
Tricium has a linter that complains if 'override final' appears in a
class method prototype since it's redundant. I encountered this linter
in a previous commit to wlan::Device and am fixing the rest of them in
wlan with this change.

Test: None necessary. Pure syntactic change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/428443
Original-Revision: 3686ae27dc0858b5b9b67780a0562dbc5305b4db
GitOrigin-RevId: 89cf065a3c1438dcb967807e3cf665d4b35deff9
Change-Id: I67102347104750aa351db530d4bf7e4d88ea03f4