[roll] Roll fuchsia [usb] Support multiple configurations for usb-peripheral mode
This patch adds support for multiple configurations bound concurrently
in peripheral mode.
It also adds support for another Realtek adapter with a proprietary
interface on the first configuration and a CDC Ethernet interface
on the second configuration. A CDC Ethernet interface is used in the
virtual-bus for both configurations due to the lack of a specification
to emulate the proprietary interface.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/428434
Original-Revision: 37df64aed0203248652eec6f0471fa14601e51b2
GitOrigin-RevId: 6c72f232c355263100fb58f61305ce03c9d59a9b
Change-Id: Ib3f3612b231b4cb2c95718300f122efd1972ca29