[roll] Roll fuchsia [crash_reports] Add a SnapshotManager
Many reports end up containing duplicate snapshots  because snapshot
requests are pooled. This is inefficeint when pool sizes are
large. To reduce the effect of large pools, reports will contain a
reference to a snapshot and the snapshot is shared between all reports
that are filed within a window of 5 seconds of one another.

One of the side effects of the previous implmentation of pooling was
that requests for a snapshot could timeout even after the snapshot had
already been collected if sycnhronous uploading took sufficiently long.
This implmentation aims to avoid that by not relying on fit::bridges to
advance the promise chain in the event of a timeout and instead checks
whether a snapshot has been collected before determining whether a
timeout has occurred.

Original-Bug: 59514

TESTED=`fx test crash-reports-tests`

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427617
Original-Revision: 5a84c05d91cb229f945fa819f01a47f36c5fa7c1
GitOrigin-RevId: cfb5d6310be2a3e21b10e3890518387cb18590a1
Change-Id: I5049da78307f00ea041dc55bf6a2ed59cb2ad44f