[roll] Roll fuchsia [time] Wait for the first request before polling for time
This is a temporary hack done instead of checking for the network to
become available. Timekeeper doesn't start making requests for time
samples until it detects that network is available. This uses the first
fidl request as a signal that the network is ready. This is done
because the network check logic is currently unstable. Once it is
stable the logic should be moved to time source and this hack removed.

Testing: fx test httpsdate-time-source-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427354
Original-Revision: 3a78294367afa2dacd4f6165b1a6e4264c0ba81c
GitOrigin-RevId: 9b2e676cc360a9d7c0889922c8ba055f71dec808
Change-Id: I59f590a8c1682e4314284ce6ac6c56169cc83ec6