[roll] Roll fuchsia [e2e] Remove can_browse_web_test
can_browse_web_test has been broken due to workstation product changes
and not running on CI builds for a considerable amount of time. Given
that a path forward for fixing the test or otherwise exercising
ChromeDriver within the open-source project doesn't seem to be appearing
soon, we are removing the test. The ChromeDriver support in the testing
framework continues to be exercised in internal tests.

Fixes: 58231
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/425976
Original-Revision: b3d3d3655f82ea2b4c0245699b3377c67d21b927
GitOrigin-RevId: 253b10dae78c43e6276baf375d5529766645f4b3
Change-Id: I21ea61eda3dc9345c5d2869b86cca2caa10c264f