[roll] Roll fuchsia [rust][fasync] Fix memory leak in the async executor
Use a weak reference to the task from the waker, to prevent Arc-cycles.
Strong refs are stored in the ready task queue and a new active_tasks
map, both of which are evicted at the end of execution to prevent
leaks.

Previously, tasks that were not completed were destroyed whenever
the last waker to it disappeared, which some code depended on.
Now that tasks are owned and have clearly defined lifetimes,
there are two types of breakages that can occur. Previous
erroneous behavior that can cause a breakage with this change:

1. Removing the last waker before the task completed ensured that
   a resource that it holds is dropped, which unblocked some
   other event in the program.
2. Never removing the waker, due to a leak, implied that a non-completed
   task never got dropped. If dropping the resource in specific state panics,
   this could hide a latent teardown issue.

Two tests of category (2) in fatfs were found and disabled. They
both panicked in drop due to an invariant assertion.

Original-Bug: 56138
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/424454
Original-Revision: ddc4f44799534a5d7e2386f5783593574e3917e6
GitOrigin-RevId: c1f2635b7b96f299d006565df461ded789c14544
Change-Id: I85edc52af4f7eee4c887ef2514278bdb4665619a