[roll] Roll fuchsia [io][conformance] Simplifies test setup with reusable FIDL setup method
Replaces overly-specific FIDL methods like GetDirectoryWithEmptyFile
with a general-purpose GetDirectory method. This method lets you provide
the structure of the entire filesystem, and will initialise it to that
state. The previous scattering of methods were each specific to a
particular test case, but this new method is reusable for any test
layout we want.

Will extend in a follow-up CL to allow adding VMO files and remote
directories to the initial state as well.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/428601
Original-Revision: 1e5401f3c4a72abc8089bba3956ac7e694f3a501
GitOrigin-RevId: 1ae09cb5c352cee5893679fae66c58a04b1769bc
Change-Id: Ibebcc9ee232834d6538f61b378914b316b8f3df1