[roll] Roll fuchsia [fx][gerrit-submit] Initial version of "fx gerrit-submit".
This tool allows chains of CLs to be automatically submitted via Gerrit
by specifying their Change-IDs.

For example, to submit three changes with Gerrit Change-IDs 101, 102,
and 103, running the command:

   fx gerrit-submit 103

The tool will:

  1. Add a CQ+1 vote to all the CLs, to start testing them.
  2. Add a CQ+2 vote for the first CL, and wait for it to be submitted.
  3. Add a CQ+2 vote for the second CL, and wait for it to be submitted

and so on.

Adding a CQ+1 to every CL at the beginning speeds up submission: CQ
won't need to re-test intermediate CLs if they are not modified in the
meantime.

If any CL is not ready to submit (for example, it is missing a vote, or
has unresolved comments), the tool will abort early.

By default, the tool will use the "fuchsia-review.googlesource.com"
Gerrit instances. Other instances can be specified using the "--host"
parameter:

   fx gerrit-submit --host myteam-review.googlesource.com <Change ID>

We use Chromium's "gerrit_util.py" as a base for making connections to
Gerrit and associated authentication, albeit with significant code
modernization and stripping out unnecessary functionality. See notes in
OSRB bug fxb/55883.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/425560
Original-Revision: 96f31919297301615ff57f8304f6b5b98f6b60ee
GitOrigin-RevId: 2e5d49f4869623c0596f4256621df49cd25a368f
Change-Id: Ibf416e5726ada7f051c3ce68b6845aa1bf7a3b14